Reading validation

+++ Atmosphere.BE/Atmosphere.Application/Readings/Commands/CreateReading.cs
@@ -9,7 +9,7 @@ using MediatR;
 public class CreateReading : IRequest<Reading>
 {
     public Guid DeviceId { get; init; }
-    public double Value { get; init; }
+    public decimal Value { get; init; }
     public DateTime Timestamp { get; init; }
     public ReadingType Type { get; init; }
 }
